pyproject.toml
[project]
name = "northwind-analytics-dbt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core>=1.11.2",
    "dbt-postgres>=1.10.0",
    "psycopg2-binary>=2.9.11",
]
